fix(argus): move case brief rendering behind client boundary [claude]
